% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tokened.R
\name{check_tokened}
\alias{check_tokened}
\alias{check_tokened.list}
\alias{check_tokened.corpus_list}
\alias{check_tokened.character}
\alias{check_tokened.VCorpus}
\alias{check_tokened.default}
\title{Check Tokenization for list}
\usage{
check_tokened(x)

\method{check_tokened}{list}(x)

\method{check_tokened}{corpus_list}(x)

\method{check_tokened}{character}(x)

\method{check_tokened}{VCorpus}(x)

\method{check_tokened}{default}(x)
}
\arguments{
\item{x}{An object containing documents}
}
\value{
A logical vector
}
\description{
This functions check if documents into a corpora are stored as single string
or are tokened
}
\examples{
data(liu_corpus)
check_token(liu_corpus)

}
