% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_for_tm.R
\name{corpora2list}
\alias{corpora2list}
\alias{list2corpora}
\title{Corpora into list and back to corpora}
\usage{
corpora2list(corpora)

list2corpora(corpora_list)
}
\arguments{
\item{corpora}{A \code{\link[tm]{Vcorpus}}}

\item{corpora_list}{A \code{corpora_list} object}
}
\value{
\code{corpora2list} returns a List of documents with metadata stored
        as attributes, i.e. document-specific metadata are stored in a list
        attribute named "metadata" of length equal of the lenght of the
        corpora; the dmeta corpora-specific metadata are stored in an
        attribute named "dmeta".

\code{list2corpora} returns a \code{\link[tm]{VCorpus}}.
}
\description{
These functions are useful to create a list of documents from a
\code{\link[tm]{VCorpus}} object (\code{corpora2list}) saving all the
information to came back to the original structure (\code{list2corpora}).
}
\details{
The aims is to provide a suitable input to function optimizated to
used native list objects such as \code{\link{list_to_dtm}} to manage the
content of the documents, and next to come back to the original structure
of \code{\link[tm]{VCorpus}}.
}
\section{Functions}{
\itemize{
\item \code{corpora2list}: Create a \code{corpora_list} object from a corpora.

\item \code{list2corpora}: Recreate a corpora from a \code{corpora_list}
object.
}}

\examples{

data(liu_4h28)
corpora  <- data2corpora(liu_4h28)
doc_list <- corpora2list(corpora)

corpora_back <- list2corpora(doc_list)
}
