% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costumer.R
\docType{package}
\name{costumer}
\alias{costumer}
\alias{costumer-package}
\title{costumer: COmprehensive Searches ThroUgh Machine learning for systEmatic Reviews}
\description{
The aims of this package is to provide the data, the functions,
scripts for the analyses and the documentation (report) within the
relative templates for the paper *Building Comprehensive Searches
including PubMed and ClinicalTrials.gov Through a Machine Learning
Approach for Systematic Reviews*.
}
