% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_train.R
\name{create_train}
\alias{create_train}
\title{Create training data structure}
\usage{
create_train(data_df, name_df = NULL)
}
\arguments{
\item{data_df}{A data frame which is expected to have (at least) the six
columns final (integer 0/1 or factror), id, title, authors and
abstract (character vectors) and year (integer).}

\item{name_df}{Character vector representing the "name" of data (to be used
       as suffix in IDs).
#'}
}
\value{
A \code{\link[tm]{VCorpus}} data structure with metadata for
        classification, id, year and authors.
}
\description{
Training data use the \pkg{tm} data structure.
}
\examples{
create_train(liu_4h28)
create_train(liu_4h28, 'Liu')
}
