% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other2stm.R
\name{dns2stm}
\alias{dns2stm}
\alias{as.simple_triplet_matrix.data.frame}
\title{Dense matrix to Simple Triplet-Matrix}
\usage{
dns2stm(dns, memory = NULL)

\method{as.simple_triplet_matrix}{data.frame}(df, memory = NULL,
  force = FALSE)
}
\arguments{
\item{dns}{a full dense \code{matrix} with rownames as documents and
colnames as tokens}

\item{memory}{(num) the maximum ammount of memory (RAM + virtual) the sistem
is required to reserve for the proces. Default is to use standard R option
and so provide "memory.limit()"}
}
\value{
an object of classes
        \code{link[tm]{DocumentTermMatrix}} and
        \code{link[slma]{simple_triplet_matrix}}
}
\description{
Dense matrix to Simple Triplet-Matrix
}
\examples{
data("liu_dtm")
liu_full_matrix <- stm2dns(liu_dtm)
identical(liu_dtm, as.simple_triplet_matrix(liu_full_matrix))
}
