% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtm_lfilter.R
\name{dtm_lfilter}
\alias{dtm_lfilter}
\title{Left filter for DTMs}
\usage{
dtm_lfilter(target, reference)
}
\arguments{
\item{target}{(\code{\link[tm]{DocumentTermMatrix}}) This is the dtm considered
for the list of documents, i.e. the returned dtm will have the exact
documents (with the same order) of this one.}

\item{reference}{(\code{\link[tm]{DocumentTermMatrix}}) This is the dtm considered
for the list of terms, i.e. the returned dtm will have the exact term (with
the same order) of this one.}
}
\value{
(\code{\link[tm]{DocumentTermMatrix}}) with weights for the terms in
        \code{reference} and documents in \code{target} DTMs.
}
\description{
Left filter for document-term matrix structure (as provided by \pkg{tm}).
I.e. return a \code{\link[tm]{DocumentTermMatrix}} for the \code{target} dtm
documents considering all (and only) the terms in the \code{reference} dtm, filling
with zeros the terms which are not present. Final order of token are the
same as in \code{target}
}
\details{
This is usefull in Machine Learning, i.e. to use a machine \code{reference}-trained
on the \code{target} dtm.
}
\examples{
library(tm)
data(acq)
data(crude)

reference <- DocumentTermMatrix(acq)
to_adjust <- DocumentTermMatrix(crude)

dtm_lfilter(reference, to_adjust)
}
\seealso{
\link{tm}, \link{slam}
}
