% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hutch.code.R
\docType{package}
\name{hutch.code}
\alias{hutch.code}
\alias{hutch.code-package}
\title{hutch.code: Highly Unbalance Text Classification in Healthcare}
\description{
The aims of this package is to provide the data, the functions, the analyses
and the documentation (report) within the relative templates for the project
about Highly Unbalance Text Classification in Healthcare.
}
