% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_for_tm.R
\name{hutch_map}
\alias{hutch_map}
\title{Transform on HUTCH corpora}
\usage{
hutch_map(corpus, FUN, ...)
}
\arguments{
\item{corpus}{A VCorpus}

\item{FUN}{a transformation function taking a text document (a character
vector) as input and returning a text document (a character vector
of the same length as the input vector).}

\item{...}{arguments to FUN}
}
\value{
A corpus with FUN applied to each document in x.
}
\description{
Interface to apply (using \pkg{purrr}) transformation functions (also
denoted as mappings) to corpora. This is a slightly modification of the
\code{link[tm]{tm_map}} function.
}
\examples{
data(liu_4h28)

corpus <- data2corpus(liu_4h28, 'id', 'Liu')
hutch_map(corpus, tolower)
}
