% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/underSam.R
\name{hutch_sampling}
\alias{hutch_sampling}
\title{Balancing of DTM}
\usage{
hutch_sampling(X, Y, type = "ROS", perc = 50, k_pos = 0, w = NULL,
  verbose = TRUE)
}
\arguments{
\item{X}{represents a DTM with the class c('DocumentTermMatrix', 'simple_triplet_matrix')}

\item{Y}{the response variable of the unbalanced dataset, should be a factor with two levels (binary)}

\item{type}{technique for balancing i.e. either Random Oversampling (ROS) or
two different ways of applying Random Undersampling  (RUS) i.e "RUS_under" type to
apply percentage of undersampling according to the majority class("percUnder") or RUS_Pos minority class("percPos")}

\item{perc}{argument for the type "RUS_under" and "RUS_Pos" only i.e. percentage of sampling of majority class depending on the type of
RUS i.e (RUS_under, RUS_Pos)}

\item{k_pos}{argument for the type "ROS" only, number of times of positve (minority) instances to be generated}

\item{w}{argument for the type "RUS_under" and "RUS_Pos only", undersampling with weighting of majotity class, if NULL sampling is done
by giving qual weights}

\item{verbose}{argument only for the type = "ROS" only. If TRUE, prints extra information}
}
\value{
if type = "RUS_Pos" or "RUS_under", value will be a list of 3 elements. The first element X will be the balanced DTM of the same class
                 as the input DTM i.e c('DocumentTermMatrix', 'simple_triplet_matrix'), the second element Y will contain the
                 response variable of the balanced data as factors and the third element will contain a vector representing the removed documents.

                 if type = "ROS", value will be a list of two elements. The first element X will be the balanced DTM of the same class
                 as the input DTM i.e c('DocumentTermMatrix', 'simple_triplet_matrix'), the second element Y will contain the
                 response variable of the balanced data as factors
}
\description{
\code{hutch_sampling} returns the balanced document term matrix using Randon Undersampling and Random Oversampling techniques
}
\details{
This function applies balancing techniques: Random Undersampling and Random Oversampling on the document term matrix
 using the functions \code{\link[unbalanced]{ubUnder}} and \code{\link[unbalanced]{ubOver}}.
}
\examples{
library(hutch.code)
library(tm)
library(unbalanced)
Y <- factor(meta(liu_corpus)$real_label)
x <- liu_dtm
exp <- hutch_sampling(stm, class, type = "RUS_Pos", perc = 50, k_pos = 0, w = NULL, verbose = FALSE)
exp$X
exp$Y
exp$id.rm

type = "ROS"
Y <- factor(meta(liu_corpus)$real_label)
x <- liu_dtm
test <- hutch_sampling(stm, class, type = "ROS", perc = 50, k_pos = 2, w = NULL, verbose = FALSE)
test$X
test$Y
}
