% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_normalizator.R
\name{idf}
\alias{idf}
\title{Inverse document-frequencies Calculator}
\usage{
idf(dtm, fun = log2, adjust = TRUE, force = FALSE)
}
\arguments{
\item{dtm}{A \code{\link[tm]{DocumentTermFrequencies}} filled with the simple
term-frequencies weights}

\item{fun}{The weighting function for the ratio of document-frequencies}

\item{adjust}{If \code{TRUE} (default) adds 1 to the denominator (i.e., the
sum of the frequencies of a term into the corpora) to avoid
the risk it happen to be zero.}

\item{force}{(lgl) if \code{TRUE} (default is \code{FALSE}) the function
works also for \code{\link[slam]{simple_triplet_matrix}}
which are not \code{\link[tm]{DocumentTermMatrix}}}
}
\value{
A numerical vector representing the inverse document-frequencies
}
\description{
The aims of this function is to compute the Inverse document-frequencies
possibly using different functions. The standard one is the \code{log2}
}
\examples{
data(liu_dtm)
idf(liu_dtm)
idf(liu_dtm, adjust = FALSE)
idf(liu_dtm, fun = log)
}
