% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{knn_cvAble}
\alias{knn_cvAble}
\title{knn_cvAble}
\format{list of 13}
\source{
\url{https://topepo.github.io/caret/using-your-own-model-in-train.html}
}
\usage{
knn_cvAble
}
\description{
k-nearest neighbour method for \pkg{caret} able to correctly manage
crossvalidation when text mining, and in particular when iDF reweighting
occurs.
}
\keyword{datasets}
