% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{liu_4h28}
\alias{liu_4h28}
\title{Liu's metanalysis positive trials and random negative ones}
\format{A data frame with 423 rows and 6 variables:

\describe{
   \item{\code{final}}{integer} gold standard classification: 1 is positive
     0 is negative.
   \item{\code{id}}{integer} ID code as retrieved from pubmed.
   \item{\code{title}}{character} Title of the record retrieved from pubmed.
   \item{\code{year}}{integer} Year of the record retrieved from pubmed.
   \item{\code{authors}}{character} Authors of the record retrieved from
     pubmed.
   \item{\code{Abstract}}{character} Abstract of the record retrieved from
     pubmed.
}}
\source{
\url{https://www.ncbi.nlm.nih.gov/pubmed/24639059}
}
\usage{
liu_4h28
}
\description{
23 Clinical Trials retrieved from pubmed records with respect to the
metanalysis by Liu,
plus 400 negative sample (taken as the first ones ordered with the "best
match" criteria on PubMED) use as negative set for training.
}
\keyword{datasets}
