% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{liu_corpus}
\alias{liu_corpus}
\title{Liu's Corpus}
\format{A \code{[tm]{VCorpus}}

\describe{
   \item{\code{content}}{named list} the text relative to each documents
   \item{\code{meta}}{list} the metadata relative to each documents
   \item{\code{dmeta}}{data.frame} the metadata relative to the corpora with the
         document relative classification
}}
\source{
\url{https://www.ncbi.nlm.nih.gov/pubmed/24639059}
}
\usage{
liu_corpus
}
\description{
This is the \code{\link[tm]{VCorpus}} representation of
\code{\link{liu_4h28}}
}
\keyword{datasets}
