% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{liu_dtm}
\alias{liu_dtm}
\title{Liu's Document-Term Matrix}
\format{Both \code{[tm]{DocumentTermMatrix}} and
        \code{[slam]{simple_triplet_matrix}}

\describe{
   \item{\code{i}}{integer vector} the document-wise (rows) coordinates of
     non-zero entries of the matrix, i.e. the lenght of i is the number of
     non-zero entries.
   \item{\code{j}}{integer vector} the term-wise (columns) coordinates of
     non-zero entries of the matrix, i.e. the lenght of j is the number of
     non-zero entries.
   \item{\code{v}}{numerical vector} the non-zero entries
     (frequencies/weight of the matrix) corresponding to the relative
     coordinates (i, j), i.e. the lenght of v is the number of non-zero
     entries.
   \item{\code{nrow}}{integer} the whole number of rows of the matrix
   \item{\code{ncol}}{integer} the whole number of columns of the matrix
   \item{\code{dimnames}}{list} a list of two character vectors representing
     the Terms/Tokens (the first one) and the name/ID of the documents (the
     second one)
}}
\source{
\url{https://www.ncbi.nlm.nih.gov/pubmed/24639059}
}
\usage{
liu_dtm
}
\description{
This is the Document-Term Matrix (with simple term frequency weights)
representation of \code{\link{liu_4h28}}
}
\keyword{datasets}
