% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowering.R
\name{lowering}
\alias{lowering}
\alias{lowering.list}
\alias{lowering.character}
\alias{lowering.VCorpus}
\alias{lowering.default}
\title{Lowering}
\usage{
lowering(corpus)

\method{lowering}{list}(corpus)

\method{lowering}{character}(corpus)

\method{lowering}{VCorpus}(corpus)

\method{lowering}{default}(corpus)
}
\arguments{
\item{corpus}{a compatible object storing documents (actually, list (and
corpus-list of (tokened) documents, character
vectors and \code{\link[tm]{VCorpus}})}
}
\value{
an object of the same class of input with documents witten with only
        lowercase letters.
}
\description{
Lowering
}
\examples{
data(liu_corpus)
lowering(c('Hello', 'World'))
lowering(liu_corpus)$content[[1]]$content
}
