% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngrams.R
\name{ngram}
\alias{ngram}
\alias{ngram.list}
\alias{ngram.VCorpus}
\alias{ngram.character}
\alias{ngram.default}
\alias{bigram}
\alias{trigram}
\title{n-Gram creators}
\usage{
ngram(corpus, n_min = 1, n_max = 2, ..., parallel = FALSE,
  ncores = parallel::detectCores() - 1)

\method{ngram}{list}(corpus, n_min = 1, n_max = 2, ..., parallel = FALSE,
  ncores = parallel::detectCores() - 1)

\method{ngram}{VCorpus}(corpus, n_min = 1, n_max = 2, ...,
  parallel = FALSE, ncores = parallel::detectCores() - 1)

\method{ngram}{character}(corpus, n_min = 1, n_max = 2, ...,
  parallel = FALSE, ncores = parallel::detectCores() - 1,
  docs_or_tokens = c("docs", "tokens"))

\method{ngram}{default}(corpus, n_min = 1, n_max = 2, ...,
  parallel = FALSE, ncores = parallel::detectCores() - 1)

bigram(corpus, ..., parallel = FALSE, ncores = parallel::detectCores() - 1)

trigram(corpus, ..., parallel = FALSE, ncores = parallel::detectCores() - 1)
}
\arguments{
\item{corpus}{a compatible object storing documents (actually, list (and
corpus-list of (tokened) documents,
character vectors and \code{\link[tm]{VCorpus}})}

\item{n_min}{(num) minimum number of words to include in the grams}

\item{n_max}{(num) maximum number of words to include into the grams}

\item{...}{further option passed to the function}

\item{parallel}{(lgl) if \code{TRUE} perform the computation in parallel
using the \code{parallel} package functionality. Default is
\code{FALSE}.}

\item{ncores}{(int) number of core to use in the parallel computation
(default is number of machine cores minus one)}

\item{docs_or_tokens}{character vector to explain if the vector is a
vector of documents (to be tokened) or is already a vector of
tokens (of a single document)}
}
\value{
an object of the same class of input (except for
        \code{character vector} input, for which the output is a \code{list})
        with documents tokenized in ngram.

(list) of character vectors containing the nGrammed documents

(list) of character vectors containing the nGrammed documents
}
\description{
The function aims is to create the ngram tokens for each document in a
corpora.

a shortcuts for \code{\link{ngram}} using
\code{n_min = n_max = 2}

a shortcuts for \code{\link{ngram}} using
\code{n_min = n_max = 3}
}
