% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonword.R
\name{nonword}
\alias{nonword}
\alias{nonword.list}
\alias{nonword.VCorpus}
\alias{nonword.character}
\alias{nonword.default}
\title{Eliminate non-words}
\usage{
nonword(corpus, numbers = FALSE, ..., pattern = NULL, replacement = " ")

\method{nonword}{list}(corpus, numbers = FALSE, ..., pattern = NULL,
  replacement = " ")

\method{nonword}{VCorpus}(corpus, numbers = FALSE, ..., pattern = NULL,
  replacement = " ")

\method{nonword}{character}(corpus, numbers = FALSE, ..., pattern = NULL,
  replacement = " ")

\method{nonword}{default}(corpus, numbers = FALSE, ..., pattern = NULL,
  replacement = " ")
}
\arguments{
\item{corpus}{a compatible object storing documents (actually, list (and
corpus-list of (tokened) documents,
character vectors and \code{\link[tm]{VCorpus}})}

\item{numbers}{(lgl) if TRUE also numbers are removed (default FALSE)}

\item{...}{Additional option}

\item{pattern}{(chr) an alternative regular expression used to remove
(i.e., to substitute with \code{replacement}) everything
that match it. Default is \code{NULL}. If not \code{NULL}
the option numbers is ignored.}

\item{replacement}{(chr) the string used to sobstitute the ones which will
be eliminated. Default is \code{' '}.}
}
\value{
an object of the same class of input with documents witten with only
        "words" retained.
}
\description{
This function's aim is to eliminate everything is not an alphanumeric
word/token from a corpora of documents. It also has an option to decide if
numbers has to be removed too. Moreover, it is possible to override both
the paramenter for the pattern identifying words and the one identifying
the replacements (default is a white space).
}
\examples{
data(liu_corpus)

nonword('hell0 w.rld')
nonword('hell0 w.rld', numbers = TRUE)                  # remove also numbers
nonword('hell0 w.rld', replacement = '*')    # use "*" instead of white space
nonword('hell0 w.rld', pattern = 'w[^\\\\s]+')     # anithing starting with "w"

nonword(liu_corpus)$content[[1]]$content # "-" removed in "anti-angiogenesis"
}
