% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reweights_test.R
\name{reweights_test}
\alias{reweights_test}
\title{Weights for Test dtm}
\usage{
reweights_test(target, original, normalize = TRUE, ..., fun = idf,
  original_weights = NULL, force = FALSE)
}
\arguments{
\item{target}{A \code{\link[tm]{DocumentTermMatrix}} to weigtht}

\item{original}{The original \code{\link[tm]{DocumentTermMatrix}} with the
simple term-frequencies weights}

\item{normalize}{(lgl) if \code{TRUE} (default) the weights are normalized
wrt documents, i.e. row_sum equals 1 for every row.}

\item{...}{Other parameters to called functions (e.g. \code{fun})}

\item{fun}{The \code{function} used to compute the original weigths. By
default \code{\link[hutch.code]{idf}}. (Suggestion is to not
change this option!). It has to admit as first argument the
orignial \code{\link[tm]{DocumentTermMatrix}}.}

\item{original_weights}{optional numerical vector of weights to be used
directly for the rewaighting of target}

\item{force}{(lgl) if \code{TRUE} (default is \code{FALSE}) the function
works also for \code{\link[slam]{simple_triplet_matrix}}
which are not \code{\link[tm]{DocumentTermMatrix}}}
}
\value{
an object of class \code{\link[tm]{DocumentTermMatrix}} and
        \code{\link[slam]{simple_triplet_matrix}} (in the order).
}
\description{
The aims of this function is to provide to a test Document-Term Matrix which
will be used for test/predict the original weights computed for the training
matrix.
}
