% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rf_cvAble}
\alias{rf_cvAble}
\title{rf_cvAble}
\format{list of 15}
\source{
\url{https://topepo.github.io/caret/using-your-own-model-in-train.html}
}
\usage{
rf_cvAble
}
\description{
random forest method for \pkg{caret} able to correctly manage
crossvalidation when text mining, and in particular when iDF reweighting
occurs.
}
\keyword{datasets}
