% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ROS3565_new}
\alias{ROS3565_new}
\alias{ROS3565_new}
\title{ROS3565_new}
\format{list of 3:

\describe{
   \item{\code{name}}{chr} the name of the function.
   \item{\code{func}}{function} the function itself.
   \item{\code{first}}{lgl} flag to decide if it happens before or after
          preprocessing.
}}
\usage{
ROS3565_new

ROS3565_new
}
\description{
Random Oversampling wit 35:65 ratio function to be used with \pkg{caret}

Random Oversampling wit 35:65 ratio function to be used with \pkg{caret}
}
\keyword{datasets}
