% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{statistics}
\alias{statistics}
\title{Calculation of different performance measures and statistical values}
\usage{
statistics(pred, truth, positive, dnn, prevalence, mode)
}
\arguments{
\item{pred}{factor containing the predicted classes}

\item{truth}{factor containing the reference classes}

\item{positive}{the first level of the factor or the level considered as positive class,
if factors are classified as positive and negative classes}

\item{dnn}{dimension names for the table}

\item{prevalence}{prevalence should be a single numeric value since we are using the binary factors}

\item{mode}{specifies either particular statistical values or everything returns all statistical values}
}
\value{
list containing two elements. First element is a list with elements: table (confusion matrix), positive (the level of positive class),
        overall (overall accuracy and other statistic values) and byClass (the values of different performance measures as specified by the argument
        mode). Second element is the value: Area under the curve
}
\description{
\code{statistics} returns the different performance measures and statical values using the function \code{\link[caret]{confusionMatrix}} and
                  the package \code{pROC} for the AUC-ROC calculation
}
\examples{
library(tm)
library(plyr)

truth <- factor(meta(liu_corpus)$real_label)
sam   <- meta(liu_corpus)$real_label
pred  <- factor(sample(sam))
statistics(pred, truth, positive = "1",  dnn = c("Prediction", "Truth"),
  prevalence = NULL, mode = "everything"
)
}
