% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem.R
\name{stem}
\alias{stem}
\alias{stem.list}
\alias{stem.character}
\alias{stem.VCorpus}
\alias{stem.default}
\title{Stem words}
\usage{
stem(corpus, language = "english")

\method{stem}{list}(corpus, language = "english")

\method{stem}{character}(corpus, language = "english")

\method{stem}{VCorpus}(corpus, language = "english")

\method{stem}{default}(corpus, language = "english")
}
\arguments{
\item{corpus}{a compatible object storing documents (actually, list (and
corpus-list of (tokened) documents,
character vectors and \code{\link[tm]{VCorpus}})}

\item{language}{(chr) language to pass to package \pkg{Snowballc} to perform
the stemming, default is 'english'.}
}
\value{
an object of the same class of input with documents written with
        all words stemmed. Names of stemmed tokens are the original un-stemed
        ones
}
\description{
Stem words
}
\examples{
sample_corpus <- list(
  c('this is', 'is a', 'a beautiful', 'beautiful day'),
  c('hello word')
)

}
