% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stm2other.R
\name{stm2dns}
\alias{stm2dns}
\alias{as.data.frame.simple_triplet_matrix}
\title{Simple Triplet-Matrix to Dense matrix}
\usage{
stm2dns(dtm, memory = NULL)

\method{as.data.frame}{simple_triplet_matrix}(dtm, memory = NULL)
}
\arguments{
\item{dtm}{an object of simple_triplet_matrix class}

\item{memory}{(num) the maximum ammount of memory (RAM + virtual) the sistem
is required to reserve for the proces. Default is to use standard R option
and so provide "memory.limit()"}
}
\value{
\code{\link{stm2dns}} returns a full dense \code{matrix}

\code{\link{as.data.frame.simple_triplet_matrix}}
         returns a \code{data.frame}
}
\description{
Simple Triplet-Matrix to Dense matrix
}
\examples{
data("liu_dtm")
liu_full_matrix <- stm2dns(liu_dtm)
liu_df <- as.data.frame(liu_dtm)
}
