% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_normalizator.R
\name{tf_normalizator}
\alias{tf_normalizator}
\title{Term Frequencies Normalizator for Document-Term Matrices}
\usage{
tf_normalizator(dtm, force = FALSE)
}
\arguments{
\item{dtm}{A \code{\link[tm]{DocumentTermFrequencies}} filled with the simple
term-frequencies weights}

\item{force}{(lgl) if \code{TRUE} (default is \code{FALSE}) the function
works also for \code{\link[slam]{simple_triplet_matrix}}
which are not \code{\link[tm]{DocumentTermMatrix}}}
}
\value{
A \code{\link[tm]{DocumentTermFrequencies}} for wich the sum of the
        frequencies of each document is 1 (or 0 if it is an empty one)
}
\description{
The aims of this function is to normalize the term frequencies in a
document-term frequencies, i.e. each frequencies of each documents in the
matrix is divided by the sum of all the frequencies of the terms in the
document.
a
}
\examples{
data(liu_dtm)
tf_normalizator(liu_dtm)
}
