% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfidf4stm.R
\name{tfidf4dtm}
\alias{tfidf4dtm}
\title{Tf-IDf for DTM}
\usage{
tfidf4dtm(dtm, normalize = TRUE, fun = log2, adjust = TRUE,
  force = FALSE)
}
\arguments{
\item{dtm}{an object of class \code{\link[tm]{DocumentTermMatrix}}
rappresenting a DTM, in particular, with documents as row.}

\item{normalize}{(lgl) if \code{TRUE} (default) the weights are normalized
wrt documents, i.e. row_sum equals 1 for every row.}

\item{fun}{The weighting function for the ratio of document-frequencies.}

\item{adjust}{If \code{TRUE} (default) adds 1 to the denominator (i.e., the
sum of the frequencies of a term into the corpora) to avoid
the risk it happen to be zero.}

\item{force}{(lgl) if \code{TRUE} (default is \code{FALSE}) the function
works also for \code{\link[slam]{simple_triplet_matrix}}
which are not \code{\link[tm]{DocumentTermMatrix}}}
}
\value{
an object of class \code{\link[tm]{DocumentTermMatrix}} and
        \code{\link[slam]{simple_triplet_matrix}} (in the order).
}
\description{
Tf-IDf for DTM
}
