% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tok_token.R
\name{tok_token}
\alias{tok_token}
\title{Tokenizator of token}
\usage{
tok_token(doc)
}
\arguments{
\item{doc}{(chr) A character vector representing a tokenized tocument}
}
\value{
(list) of character vector each element representing
        the word-components of each original token
}
\description{
The function's aim is to access tokens at their single-words component level
maintaining the information of the original token structure
}
\examples{
tokened_document <- c('this is', 'is a', 'a beautiful', 'beautiful day')

tok_token(tokened_document)
}
