% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_for_tm.R
\name{v_assign_meta}
\alias{v_assign_meta}
\title{Assign a vector of metadata to a corpus}
\usage{
v_assign_meta(corpus, meta_name, meta_content)
}
\arguments{
\item{corpus}{A VCorpus}

\item{meta_name}{A character vector representing the name of meta to consider}

\item{meta_content}{A vector or a list of data of length of corpus containing
the corresponding meta data content to assign each one}
}
\value{
A VCorpus (invisible) with metadata updated
}
\description{
This function provide the possibility to assign a vector of metadata (one
each document) to a copus in one command.
To assign a single metadata to a single document or to the whole corpuse
use the standard \code{\link[tm]{meta}}.
}
\examples{
library(hutch.code)
data(liu_4h28)

corpus <- create_train(data_df = liu_4h28, name_df = 'Liu')
corpus <- assign_meta(corpus, 'author', liu_4h28$authors)
}
