% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/white.R
\name{white}
\alias{white}
\alias{white.list}
\alias{white.character}
\alias{white.VCorpus}
\alias{white.default}
\title{Strip White Spaces}
\usage{
white(corpus, ..., pattern = "\\\\s+")

\method{white}{list}(corpus, ..., pattern = NULL)

\method{white}{character}(corpus, ..., pattern = NULL)

\method{white}{VCorpus}(corpus, ..., pattern = NULL)

\method{white}{default}(corpus, ..., pattern = NULL)
}
\arguments{
\item{corpus}{a compatible object storing documents (actually, list (and
\code{\link[hutch.code]{corpus_list}}) of (tokened) documents,
character vectors and \code{\link[tm]{VCorpus}})}

\item{...}{Other paramenter}

\item{pattern}{(chr) A regular expression to be use for detection of
withespace. If \code{NULL} (default), \code{\\s+} is used.}
}
\value{
an object of the same class of input with documents witten with
        trimmed whitespaces.
}
\description{
The function \code{white} collapse all multiple "spaces" in a single space.
By default the function identifies a white space by \code{\\s+} which is
a shortcut for \code{[^[:space:]]} i.e. tab, newline, vertical tab, form
feed, carriage return, space and possibly other locale-dependent characters.
There is option to override the pattern to be use to identify white spaces.
}
\examples{
data(liu_corpus)

white(c(' one  two   three    '))
white(liu_corpus)
}
